✨ feat: expand CPU architecture to 32-bit instruction and register model
- Change opcode from 16-bit to 32-bit format
- Extend general-purpose registers from V0~VF to R0~R31
- Replace VF with FLAG_REG (R[31]) for carry/borrow flags
- Update opcode parsing and decoding logic for 32-bit layout
- Modify PC, I, stack, and memory access to support 32-bit architecture
- Maintain compatibility in primary opcode table structure

Related to future test ROM support with 32-bit instruction format

+++ include/core/opcode_table.hpp
@@ -14,7 +14,7 @@ class Chip8; // 전방 선언 (헤더에서 Chip8 전체 정의 불필요)
 
 namespace OpcodeTable {
 
-    using OpcodeHandler = std::function<void(Chip8&, uint16_t)>;
+    using OpcodeHandler = std::function<void(Chip8&, uint32_t)>; // (change)
 
     // 명령어 0x0000 ~ 0xFFFF 중, 상위 4비트 또는 특정 패턴으로 구분하여 핸들러를 매핑합니다.
     extern std::array<OpcodeHandler, 16> primary_table; // 예: 0x1000 => Jump
@@ -25,6 +25,6 @@ namespace OpcodeTable {
      */
     void Initialize();
 
-    void Execute(Chip8& chip8, uint16_t opcode);
+    void Execute(Chip8& chip8, uint32_t opcode); // (change)
 
 } // namespace OpcodeTable
